emergency stop correction to function
Added a new routine check_for_emergency_stop which positions the stepper at either the open or close position depending upon whether it was opening or closing when the ES was activated. This ensures  the stepper does not over run when an open or close command is executed after ES. Needs test.